test confirmation pop up

+++ Controller/controleReservation.java
@@ -39,23 +39,26 @@ public class controleReservation implements ActionListener{
 				JOptionPane.showMessageDialog(mainView, "Vous n'avez pas choisis de reservation.","Pas de reservation",JOptionPane.WARNING_MESSAGE);
 			}else{
 				// Ajouter les réservations
-				
-				double prix = compte.getuser().getQuotient().getTotal();
-				double total_prix = prix * list_resa.size();
-				double manquant = total_prix - compte.getSolde();
-				if(compte.getSolde() - total_prix >= 0){
-					for(String s : list_resa){
-						Reservation r = new Reservation(page.getEnfant(), s, compte);
-						compte.addResa(r);
+				JOptionPane.showConfirmDialog(mainView,"Voulez vous vraiment confirmer cette réservation?","Confirmation",JOptionPane.YES_NO_OPTION); {
+						
+						double prix = compte.getuser().getQuotient().getTotal();
+						double total_prix = prix * list_resa.size();
+						double manquant = total_prix - compte.getSolde();
+						if(compte.getSolde() - total_prix >= 0){
+							for(String s : list_resa){
+								Reservation r = new Reservation(page.getEnfant(), s, compte);
+								compte.addResa(r);
+							}
+						compte.debiter(total_prix);
+					}else {
+						JOptionPane.showMessageDialog(mainView, "Vous n'avez pas assez de solde pour faire ces reservations: il vous manque "+new DecimalFormat(".##").format(manquant),"Solde insuffisant",JOptionPane.WARNING_MESSAGE);
 					}
-					compte.debiter(total_prix);
-				}else {
-					JOptionPane.showMessageDialog(mainView, "Vous n'avez pas assez de solde pour faire ces reservations: il vous manque "+new DecimalFormat(".##").format(manquant),"Solde insuffisant",JOptionPane.WARNING_MESSAGE);
+					page.showResa();
+					page.deleteDate();
+					mainView.validate();
+					}
 				}
-				page.showResa();
-				page.deleteDate();
-				mainView.validate();
-			}
+			
 		}
 		else if(pressed.getName() == "retour"){
 			mainView.getContentPane().removeAll();
